+++ virtio/9p.c
@@ -1382,7 +1382,7 @@ static size_t get_config_size(struct kvm *kvm, void *dev)
 	return p9dev->config_size;
 }
 
-static u32 get_host_features(struct kvm *kvm, void *dev)
+static u64 get_host_features(struct kvm *kvm, void *dev)
 {
 	return 1 << VIRTIO_9P_MOUNT_TAG;
 }
