+++ virtio/blk.c
@@ -169,7 +169,7 @@ static size_t get_config_size(struct kvm *kvm, void *dev)
 	return sizeof(bdev->blk_config);
 }
 
-static u32 get_host_features(struct kvm *kvm, void *dev)
+static u64 get_host_features(struct kvm *kvm, void *dev)
 {
 	struct blk_dev *bdev = dev;
